+++ Jitzu.Shell/Infrastructure/Logging/ConsoleEx.cs
@@ -1,10 +1,10 @@
-﻿namespace Jitzu.Interpreter.Infrastructure.Logging;
+namespace Jitzu.Shell.Infrastructure.Logging;
 
 public static class ConsoleEx
 {
     public static void ConfigureOutput()
     {
-        if (!Console.IsOutputRedirected && !Console.IsInputRedirected) 
+        if (!Console.IsOutputRedirected && !Console.IsInputRedirected)
             return;
 
         var stdOut = Console.OpenStandardOutput();